+++ src/js/Bounds.d.ts
@@ -5,13 +5,15 @@
  *
  * @author   Ikaros Kappler
  * @date     2020-05-11
- * @version  1.0.0
+ * @modified 2020-10-30 Added the static computeFromVertices function.
+ * @version  1.1.0
  *
  * @file Bopunds
  * @fileoverview A simple bounds class implementing IBounds.
  * @public
  **/
 import { XYCoords, IBounds, XYDimension } from "./interfaces";
+import { Vertex } from "./Vertex";
 export declare class Bounds implements IBounds, XYDimension {
     /**
      * @member {XYCoords}
@@ -50,4 +52,16 @@ export declare class Bounds implements IBounds, XYDimension {
      * @param {XYCoords} max - The max values (x,y) as a XYCoords tuple.
      **/
     constructor(min: XYCoords, max: XYCoords);
+    /**
+     * Compute the minimal bounding box for a given set of vertices.
+     *
+     * An empty vertex array will return an empty bounding box located at (0,0).
+     *
+     * @static
+     * @method computeFromVertices
+     * @memberof Bounds
+     * @param {Array<Vertex>} vertices - The set of vertices you want to get the bounding box for.
+     * @return The minimal Bounds for the given vertices.
+     **/
+    static computeFromVertices(vertices: Array<Vertex>): Bounds;
 }