+++ src/js/Bounds.js
@@ -6,13 +6,15 @@
  *
  * @author   Ikaros Kappler
  * @date     2020-05-11
- * @version  1.0.0
+ * @modified 2020-10-30 Added the static computeFromVertices function.
+ * @version  1.1.0
  *
  * @file Bopunds
  * @fileoverview A simple bounds class implementing IBounds.
  * @public
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
+var Vertex_1 = require("./Vertex");
 var Bounds = /** @class */ (function () {
     /**
      * The constructor.
@@ -28,6 +30,35 @@ var Bounds = /** @class */ (function () {
         this.width = max.x - min.x;
         this.height = max.y - min.y;
     }
+    ;
+    /**
+     * Compute the minimal bounding box for a given set of vertices.
+     *
+     * An empty vertex array will return an empty bounding box located at (0,0).
+     *
+     * @static
+     * @method computeFromVertices
+     * @memberof Bounds
+     * @param {Array<Vertex>} vertices - The set of vertices you want to get the bounding box for.
+     * @return The minimal Bounds for the given vertices.
+     **/
+    Bounds.computeFromVertices = function (vertices) {
+        if (vertices.length == 0)
+            return new Bounds(new Vertex_1.Vertex(0, 0), new Vertex_1.Vertex(0, 0));
+        var xMin = vertices[0].x;
+        var xMax = vertices[0].x;
+        var yMin = vertices[0].y;
+        var yMax = vertices[0].y;
+        var vert;
+        for (var i in vertices) {
+            vert = vertices[i];
+            xMin = Math.min(xMin, vert.x);
+            xMax = Math.max(xMax, vert.x);
+            yMin = Math.min(yMin, vert.y);
+            yMax = Math.max(yMax, vert.y);
+        }
+        return new Bounds(new Vertex_1.Vertex(xMin, xMax), new Vertex_1.Vertex(yMin, yMax));
+    };
     ;
     return Bounds;
 }()); // END class bounds