+++ src/ts/Bounds.ts
@@ -5,7 +5,8 @@
  * 
  * @author   Ikaros Kappler
  * @date     2020-05-11
- * @version  1.0.0
+ * @modified 2020-10-30 Added the static computeFromVertices function.
+ * @version  1.1.0
  *
  * @file Bopunds
  * @fileoverview A simple bounds class implementing IBounds.
@@ -13,6 +14,7 @@
  **/
 
 import { XYCoords, IBounds, XYDimension } from "./interfaces";
+import { Vertex } from "./Vertex";
 
 export class Bounds implements IBounds, XYDimension {
 
@@ -63,6 +65,38 @@ export class Bounds implements IBounds, XYDimension {
 	this.max = max;
 	this.width = max.x-min.x;
 	this.height = max.y-min.y;
+    };
+
+    /**
+     * Compute the minimal bounding box for a given set of vertices.
+     *
+     * An empty vertex array will return an empty bounding box located at (0,0).
+     *
+     * @static
+     * @method computeFromVertices
+     * @memberof Bounds
+     * @param {Array<Vertex>} vertices - The set of vertices you want to get the bounding box for.
+     * @return The minimal Bounds for the given vertices.
+     **/
+    static computeFromVertices( vertices:Array<Vertex> ) : Bounds {
+	if( vertices.length == 0 )
+	    return new Bounds( new Vertex(0, 0), new Vertex(0, 0) );
+
+	let xMin = vertices[0].x;
+	let xMax = vertices[0].x;
+	let yMin = vertices[0].y;
+	let yMax = vertices[0].y;
+
+	let vert:Vertex;
+	for( var i in vertices ) {
+	    vert = vertices[ i ];
+	    xMin = Math.min( xMin, vert.x );
+	    xMax = Math.max( xMax, vert.x );
+	    yMin = Math.min( yMin, vert.y );
+	    yMax = Math.max( yMax, vert.y );
+	}
+
+	return new Bounds( new Vertex(xMin, xMax), new Vertex(yMin, yMax) );
     };
 
 } // END class bounds