+++ changelog.md
@@ -1,8 +1,9 @@
 
 
 # Changelog
-* 2020-10-27
-  * Added the draw.svgPath(...) function.
+* 2020-10-30
+  * Added the static Bounds.computeFromVertices(Vertex[]) function.
+  * Added the `Polygon.addVert` function.
 * 2020-10-23
   * Added the fillArray and fillMatrix helper function (like in lodash).
   * Added the `Matrix`, `Interval` and `IndexPair` interfaces to the typescript definitions.